requirements.txt
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.0.350
langchain-openai>=0.0.2
youtube-transcript-api>=0.6.1
pydantic>=2.4.2
openai>=1.3.0
httpx>=0.25.0
python-multipart>=0.0.6
markdown>=3.5.1
APScheduler>=3.10.4
aiosmtplib>=2.0.2